Refactoring tests in their own package. First steps in test for superclass and subclasses.
